requirements.txt
streamlit
python-dotenv
requests
pydub
pyttsx3
transformers
torch
scipy